ayble/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
environ==1.0
importlib-metadata==6.0.0
Markdown==3.4.1
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.11.0
